requirements_minimal.txt
# ── Core AI framework ──────────────────────────────────────────────────────────
# NOTE: crewai requires Python >=3.10,<3.14 — use Python 3.12 venv
crewai==0.130.0
crewai-tools==0.47.1

# ── API server ─────────────────────────────────────────────────────────────────
fastapi==0.110.3
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# ── Queue worker (Celery + Redis) ──────────────────────────────────────────────
celery>=5.3.6
redis>=5.0.4

# ── Database ───────────────────────────────────────────────────────────────────
sqlalchemy>=2.0.30

# ── PDF loading ────────────────────────────────────────────────────────────────
pypdf>=4.2.0
langchain-community>=0.0.38

# ── Env & utilities ────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
